requirements.txt
# UI & Dashboard
streamlit==1.32.0

# LLM & AI Processing
openai==1.13.3
python-dotenv==1.0.1

# Vector Database (Institutional Memory)
chromadb==0.4.24

# Quantitative Engine & Data Handling
pandas==2.2.1
scipy==1.12.0
numpy==1.26.4
